+++ src/main/java/com/rebook/book/repository/BookRepository.java
@@ -1,6 +1,6 @@
 package com.rebook.book.repository;
 
-import com.rebook.book.domain.BookEntity;
+import com.rebook.book.domain.entity.BookEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookRepository extends JpaRepository<BookEntity, Long> {
