+++ src/main/java/com/rebook/book/service/BookService.java
@@ -1,6 +1,6 @@
 package com.rebook.book.service;
 
-import com.rebook.book.domain.BookEntity;
+import com.rebook.book.domain.entity.BookEntity;
 import com.rebook.book.dto.response.BookResponse;
 import com.rebook.book.dto.request.BookCreateRequest;
 import com.rebook.book.repository.BookRepository;
